+++ src/link/error/index.ts
@@ -11,13 +11,29 @@ import {
 import { ApolloLink } from "@apollo/client/link";
 
 export declare namespace ErrorLink {
-  /**
-   * Callback to be triggered when an error occurs within the link stack.
-   */
+  // Using a different namespace name to avoid clash with
+  // `ApolloLink.DocumentationTypes`
+  export namespace ErrorLinkDocumentationTypes {
+    /**
+     * Callback that is called by `ErrorLink` when an error occurs from a
+     * downstream link in link chain.
+     *
+     * @param options - The options object provided by `ErrorLink` to the error
+     * handler when an error occurs.
+     */
+    export function ErrorHandler(
+      options: ErrorHandlerOptions
+    ): Observable<ApolloLink.Result> | void;
+  }
+
+  /** {@inheritDoc @apollo/client/link/error!ErrorLink.ErrorLinkDocumentationTypes.ErrorHandler:function(1)} */
   export interface ErrorHandler {
     (options: ErrorHandlerOptions): Observable<ApolloLink.Result> | void;
   }
 
+  /**
+   * The object provided to the `ErrorHandler` callback function.
+   */
   export interface ErrorHandlerOptions {
     /**
      * The error that occurred during the operation execution. This can be a
@@ -46,24 +62,65 @@ export declare namespace ErrorLink {
   }
 }
 
-// For backwards compatibility.
-export import ErrorHandler = ErrorLink.ErrorHandler;
-
 /**
  * @deprecated
  * Use `ErrorLink` from `@apollo/client/link/error` instead.
  */
-export function onError(errorHandler: ErrorHandler) {
+export function onError(errorHandler: ErrorLink.ErrorHandler) {
   return new ErrorLink(errorHandler);
 }
 
+/**
+ * Use the `ErrorLink` to perform custom logic when a [GraphQL or network error](https://apollographql.com/docs/react/data/error-handling)
+ * occurs.
+ *
+ * @remarks
+ *
+ * This link is used after the GraphQL operation completes and execution is
+ * moving back up your [link chain](https://apollographql.com/docs/react/introduction#handling-a-response). The `errorHandler` function should
+ * not return a value unless you want to [retry the operation](https://apollographql.com/docs/react/data/error-handling#retrying-operations).
+ *
+ * For more information on the types of errors that might be encountered, see
+ * the guide on [error handling](https://apollographql.com/docs/react/data/error-handling).
+ *
+ * @example
+ *
+ * ```ts
+ * import { ErrorLink } from "@apollo/client/link/error";
+ * import {
+ *   CombinedGraphQLErrors,
+ *   CombinedProtocolErrors,
+ * } from "@apollo/client/errors";
+ *
+ * // Log any GraphQL errors, protocol errors, or network error that occurred
+ * const errorLink = new ErrorLink(({ error, operation }) => {
+ *   if (CombinedGraphQLErrors.is(error)) {
+ *     error.errors.forEach(({ message, locations, path }) =>
+ *       console.log(
+ *         `[GraphQL error]: Message: ${message}, Location: ${locations}, Path: ${path}`
+ *       )
+ *     );
+ *   } else if (CombinedProtocolErrors.is(error)) {
+ *     error.errors.forEach(({ message, extensions }) =>
+ *       console.log(
+ *         `[Protocol error]: Message: ${message}, Extensions: ${JSON.stringify(
+ *           extensions
+ *         )}`
+ *       )
+ *     );
+ *   } else {
+ *     console.error(`[Network error]: ${error}`);
+ *   }
+ * });
+ * ```
+ */
 export class ErrorLink extends ApolloLink {
   constructor(errorHandler: ErrorLink.ErrorHandler) {
     super((operation, forward) => {
       return new Observable((observer) => {
         let sub: Subscription | undefined;
         let retriedSub: Subscription | undefined;
-        let retriedResult: ReturnType<ErrorHandler>;
+        let retriedResult: ReturnType<ErrorLink.ErrorHandler>;
 
         try {
           sub = forward(operation).subscribe({