+++ src/link/subscriptions/index.ts
@@ -47,6 +47,29 @@ function isLikeErrorEvent(err: unknown): err is Event {
   return isNonNullObject(err) && err.target?.readyState === WebSocket.CLOSED;
 }
 
+/**
+ * The `GraphQLWsLink` is a terminating link sends GraphQL operations over a
+ * WebSocket connection using the [`graphql-ws`](https://www.npmjs.com/package/graphql-ws) library. It's used most
+ * commonly with GraphQL [subscriptions](https://apollographql.com/docs/react/data/subscriptions/),
+ *
+ * > [!NOTE]
+ * > This link works with the `graphql-ws` library. If your server uses
+ * > the deprecated `subscriptions-transport-ws` library, use the deprecated
+ * > [`WebSocketLink`](https://apollographql.com/docs/react/api/link/apollo-link-ws) link instead.
+ *
+ * @example
+ *
+ * ```ts
+ * import { GraphQLWsLink } from "@apollo/client/link/subscriptions";
+ * import { createClient } from "graphql-ws";
+ *
+ * const link = new GraphQLWsLink(
+ *   createClient({
+ *     url: "ws://localhost:3000/subscriptions",
+ *   })
+ * );
+ * ```
+ */
 export class GraphQLWsLink extends ApolloLink {
   constructor(public readonly client: Client) {
     super();